-Transition component is now appling the style on the child component -extracted logic for generating the animationStyle so that it can be reused in renderComponent, as well as the Transition component
