+++ docs/src/examples/components/Transition/Types/TransitionExample.shorthand.tsx
@@ -3,12 +3,11 @@ import { Transition, Icon, Button } from '@stardust-ui/react'
 
 const TransitionExampleShorthand = () => (
   <div>
-    <Transition animationName="colorChanger" content="Hello from Transition!" />
     <Transition animationName="spinner">
       <Icon name="umbrella" circular />
     </Transition>
     <Transition animationName="spinner">
-      <Button content={'I am spinning'} primary />
+      <Button content={'I am spinning'} />
     </Transition>
   </div>
 )
